disabled half load in booking.tsx
